mediatek-sdk: Make BOARD mediatek_sdk
Feed names cannot contain -.  For targets provided by a feed, the
feed name must match the BOARD variable in the target Makefile. So,
it would seem that the BOARD variable for the mediatek-sdk feed
should be mediatek_sdk, so that any profile that requires it can
specify the feed name as mediatek_sdk and end up with a working
target.