q14: Properly preserve settings during sysupgrade
During flashing the current settings are saved in sysupgrade.tar.
This file needs to be tucked away somewhere persistent, so that
on the subsequent boot, it can be copied to /, where
/lib/preinit/80_mount_root expects to find it.  In our case, we
will use the unused rootfs_data partition as our persistent storage.
The saved settings are copied there during sysupgrade, and then
moved from there to / during preinit.

SW-2293