patches: minim: Remove regdb patch for 5850-5895
This patch was added to allow operation in UNII-4 bands.  However,
we no longer want to support UNII-4 bands for R14.  Originally,
rather then remove this patch, we just removed the UNII-4 bands
from the devices auto channel selection list.  However, this is
not sufficient for preventing operation in UNII-4 as the driver
might still choose a UNII-4 channel when detecting radar.  The only
way to ensure no UNII-4 channels are used is to remove this patch

SW-3937